+++ libhls/source/hls-fmp4.c
@@ -80,9 +80,9 @@ static int mov_buffer_seek(void* param, int64_t offset)
 	return 0;
 }
 
-static uint64_t mov_buffer_tell(void* param)
+static int64_t mov_buffer_tell(void* param)
 {
-	return ((struct hls_fmp4_t*)param)->offset;
+	return (int64_t)((struct hls_fmp4_t*)param)->offset;
 }
 
 static struct mov_buffer_t s_io = {
